Add files via upload
"This is a project in which you can detect your location by using your phone number by giving your phone number and the count of numbers in your country code. In this project, we are connected with SQL, which has a table of all phone numbers and country codes. You are sending the country code and getting the country name."